backend/pyproject.toml
[tool.poetry]
name = "pixiserve"
version = "0.1.0"
description = "Self-hosted photo management with ML-powered organization"
authors = ["Bryan"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"

# Web framework
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.6"

# Database
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"

# Redis & Task Queue
redis = "^5.0.1"
celery = {extras = ["redis"], version = "^5.3.6"}

# Authentication
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"

# Storage
boto3 = "^1.34.25"
aiofiles = "^23.2.1"

# Configuration
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"

# Image Processing
pillow = ">=10.0.0"

# EXIF & Metadata
exifread = "^3.0.0"

# Geocoding
geopy = "^2.4.1"

# ML Core (CPU fallback, GPU-specific installed separately)
numpy = ">=1.24.0"
onnxruntime = ">=1.16.0"
scikit-learn = ">=1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^24.1.0"
ruff = "^0.1.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
